askbot_requirements.txt
Coffin>=0.3
Jinja2
Lamson
South>=0.7.1
akismet
beautifulsoup4
django-avatar>=2.0
django-celery>=3.0.11
django-compressor==1.2
django-countries==1.0.5
django-followit==0.0.7
django-keyedcache
django-kombu==0.9.4
django-picklefield==0.3.0
django-recaptcha-works
django-robots
django-threaded-multihost
django-tinymce==1.5.1b2
django>=1.3.1,<1.6
html5lib==0.90
longerusername
markdown2
oauth2
pyjwt
pystache==0.3.1
python-openid
pytz==2013b
sanction==0.3.1
unidecode
